requirements.txt
Django==5.1.2
psycopg[binary]==3.2.3
python-dotenv==1.0.1
Pillow==11.0.0
whitenoise==6.7.0
